RD-2 Move GameStatus model file
