+++ src/app/game/models/game-status.d.ts
@@ -1,13 +1,3 @@
-export interface GameMessage {
-   type: string;
-   data: any;
-}
-
-export interface InputMessage {
-   type: string;
-   data: any;
-}
-
 export interface GameStatus {
    isGameGoing: boolean;
    isPaused: boolean;
