+++ src/app/game/models/index.ts
@@ -1,0 +1,3 @@
+//@index('./*', (f, _) => `export { ${_.pascalCase(f.name.split('.')[0])} } from '${f.path}';`)
+export { GameStatus } from './game-status.d';
+//@endindex
